refactor: mv treeland-wallpaper to examples
Log:
