+++ examples/CMakeLists.txt
@@ -16,3 +16,4 @@ add_subdirectory(test_super_overlay_surface)
 add_subdirectory(test_monitor_active_event)
 add_subdirectory(test_primary_output)
 add_subdirectory(test_multitaskview)
+add_subdirectory(test_set_treeland_wallpaper)
