+++ examples/test_set_treeland_wallpaper/PropertyItemDelegate.qml
