+++ examples/test_set_treeland_wallpaper/Wallpaper.qml
