+++ examples/test_set_treeland_wallpaper/WallpaperSetting.qml
