+++ examples/test_set_treeland_wallpaper/personalizationmangerclient.cpp
